requirements.txt
pyvoro
numpy-stl
